Fix compilation failure on Win64 since r14020.
R=danno@chromium.org

Review URL: https://codereview.chromium.org/12989002

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@14035 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/heap.cc
@@ -3880,7 +3880,7 @@ MaybeObject* Heap::CopyCode(Code* code, Vector<byte> reloc_info) {
   Address new_addr = reinterpret_cast<HeapObject*>(result)->address();
 
   // Copy header and instructions.
-  CopyBytes(new_addr, old_addr, relocation_offset);
+  CopyBytes(new_addr, old_addr, static_cast<int>(relocation_offset));
 
   Code* new_code = Code::cast(result);
   new_code->set_relocation_info(ByteArray::cast(reloc_info_array));
